Unreal Engine Support V_0.30 - Beta support - T... - TeamMythic
Updated some Unreal Support in preparation of Farming Game for everyone to make. Also added a new scriptable object btn under Assets/GettingStarted: this houses 2 btns for the Website and the Docs website. Lastly, made a few folder deletions and renaming schemes for what fits right in regards to the current status of the open source and how we move forward